add "init" subcommand, and move most cli flags to a config file
with the init command it now takes fewer steps to get started.

for future development, it will be easier to change config options.

we had quite some cli flags before, which couldn't easily be changed without
also changing how ysco is started. with a config file, the command-line
invocation is simpler. the admin can make config changes through the web
interface too now (spartan at the moment, can be improved in the future).