when putting new binaries in place, possibly through symlinks, evaluate to absolute paths for proper comparison for equality, and fix removing them
and some refactoring with better variable names.

previous ysco binaries weren't being removed because we were comparing
the directory of the service binary (relative by default) with the
directory of the ysco binary (absolute by default), which was wrong in
two ways: absolute vs relative, and always using the service binary.